OGR-VFK: better reporing invalid features
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@24049 f0d54148-0727-0410-94bb-9a71ac55c965

+++ ogr/ogrsf_frmts/vfk/vfkreaderp.h
@@ -74,7 +74,7 @@ public:
     OGRErr         OpenFile(const char *);
     int            ReadDataBlocks();
     int            ReadDataRecords(IVFKDataBlock *);
-    long           LoadGeometry();
+    int            LoadGeometry();
     
     int            GetDataBlockCount() const { return m_nDataBlockCount; }
     IVFKDataBlock *GetDataBlock(int) const;
